renaming + extracting enum
