+++ src/components/ActNowCard/enums.js
@@ -1,0 +1,9 @@
+const selectedCard = {
+  CONNECT: "connect",
+  FOLLOW: "follow",
+  PETITION: "petition",
+  SHARE: "share",
+  ACT: "act",
+};
+
+export default selectedCard;
